added "Options / GeoIP Service" menu to select between different implementations.
It now defaults to the Maxmind GeoLite2 local offline database.
The previous implementation based on ip-api.com's free API can be used alternatively.
