+++ ServerBrowser/GeoIp/COPYRIGHT.txt
@@ -1,0 +1,1 @@
+Database and Contents Copyright (c) 2024 MaxMind, Inc.
