+++ ServerBrowser/GeoIp/README.txt
@@ -1,0 +1,1 @@
+Latitude and longitude are not precise and should not be used to identify a particular street address or household.
