+++ ServerBrowser/app.config
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
     <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
@@ -11,34 +11,44 @@
   <applicationSettings>
     <DevExpress.LookAndFeel.Design.AppSettings>
       <setting name="DefaultAppSkin" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="DefaultPalette" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="TouchUI" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="CompactUI" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="TouchScaleFactor" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="DirectX" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="RegisterUserSkins" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="RegisterBonusSkins" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="FontBehavior" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="DefaultAppFont" serializeAs="String">
-        <value></value>
+        <value>
+        </value>
       </setting>
       <setting name="DPIAwarenessMode" serializeAs="String">
         <value>PerMonitorV2</value>
@@ -50,11 +60,42 @@
   </applicationSettings>
   <runtime>
     <loadFromRemoteSources enabled="true" />
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="System.Runtime.CompilerServices.Unsafe" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-6.0.0.0" newVersion="6.0.0.0" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="System.Text.Json" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-8.0.0.4" newVersion="8.0.0.4" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.Extensions.Options" publicKeyToken="adb9793829ddae60" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-8.0.0.2" newVersion="8.0.0.2" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.Extensions.DependencyInjection.Abstractions" publicKeyToken="adb9793829ddae60" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-8.0.0.1" newVersion="8.0.0.1" />
+      </dependentAssembly>
+    </assemblyBinding>
   </runtime>
   <startup>
     <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8" />
   </startup>
   <appSettings>
     <add key="EnableWindowsFormsHighDpiAutoResizing" value="true" />
+    <add key="ClientSettingsProvider.ServiceUri" value="" />
   </appSettings>
+  <system.web>
+    <membership defaultProvider="ClientAuthenticationMembershipProvider">
+      <providers>
+        <add name="ClientAuthenticationMembershipProvider" type="System.Web.ClientServices.Providers.ClientFormsAuthenticationMembershipProvider, System.Web.Extensions, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" serviceUri="" />
+      </providers>
+    </membership>
+    <roleManager defaultProvider="ClientRoleProvider" enabled="true">
+      <providers>
+        <add name="ClientRoleProvider" type="System.Web.ClientServices.Providers.ClientRoleProvider, System.Web.Extensions, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" serviceUri="" cacheTimeout="86400" />
+      </providers>
+    </roleManager>
+  </system.web>
 </configuration>
