+++ ServerBrowser/packages.config
@@ -1,4 +1,18 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="ZeroMQ" version="4.1.0.15" targetFramework="net40-Client" />
+  <package id="MaxMind.Db" version="4.1.0" targetFramework="net48" />
+  <package id="MaxMind.GeoIP2" version="5.2.0" targetFramework="net48" />
+  <package id="Microsoft.Bcl.AsyncInterfaces" version="8.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="8.0.1" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Options" version="8.0.2" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Primitives" version="8.0.0" targetFramework="net48" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
+  <package id="System.Memory" version="4.5.5" targetFramework="net48" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net48" />
+  <package id="System.Text.Encodings.Web" version="8.0.0" targetFramework="net48" />
+  <package id="System.Text.Json" version="8.0.4" targetFramework="net48" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net48" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net48" />
+  <package id="ZeroMQ" version="4.1.0.31" targetFramework="net48" />
 </packages>
